nvim: enable creating undofiles
nvim automatically saves undo history to an undo file when
writing a buffer to a file, and restores undo history from the same
file on buffer read.
The directory where the undo file is stored is specified by 'undodir'.

I use ~/data/ as my 'undodir', as ~/data/ is the root for my backups
(data that is not in version control).