nvim: add <leader>lr for LSP restart
This is a bit quirky, but it works for me perfectly and I don't know how
to refactor it right now.

The 100ms sleep between LSP restart and reloading the file is required
for LSP to start reindex.

Not the most beautiful solution, but it will have to do for now.